Refactor: Consolidate ImagesResponse imports and remove legacy file
